refactor(signTransaction): use signature directly
### Problem Statement

This is precluding me from working around limitations related to
secp256k1 support so I have to make my own transaction object and
rewrite some of this.

### Solution

This is un-necessary re-creation of the Signature object. Doing things
this way allows spoofing but also reduces a small amount of redundant
code.

### How has this been tested?

This should be tested heavily by existing CI most likely.